refactor: move components to a different file
